# optimize CLASS_OF