`CLASS_OF` is literally everywhere.  When you call a method, the class has to be inferred from its receiver, which is done by this.

So even a small fraction of it has a very huge impact on runtime performance.  Looking at the implementation the `RBASIC(obj)` part was at the very end of the function body.  This, however, is
expected to be the most frequently-entering branch because most method calls for Fixnums etc. are special-cased in `insns.def`.

Moving what is likely to happen right in front of the function makes it easier for compilers to emit optimal codes.  This rearrangement does not speed things up at all if you use gcc 9, but for clang 10, the impact is considerable.

```
Calculating -------------------------------------
                         ours@clang-10  ours@gcc-9  master@clang-10 master@gcc-9
Optcarrot Lan_Master.nes        45.471      47.408           38.084       44.744 fps

Comparison:
             Optcarrot Lan_Master.nes
              ours@gcc-9:        47.4 fps
           ours@clang-10:        45.5 fps - 1.04x  slower
            master@gcc-9:        44.7 fps - 1.06x  slower
         master@clang-10:        38.1 fps - 1.24x  slower
```