@methodmissing 2019-10-23
Nice work on this. As an aside I would also note that this same optimization could translate well to the `TYPE` macro and `rb_type` as well. It's structurally similar, however not in as critical a path as method dispatch. That said it drives quite a few `switch` statements in core objects and bundled extensions and also the meat of the `checktype` instruction.