@shyouhei 2019-10-24
@methodmissing Yes I was aware of `rb_type`.  Didn't touch here for the sake of pull request simplicity.